Time: 4 ms (47.53%), Space: 9 MB (46.23%) - LeetHub

+++ 0326-power-of-three/0326-power-of-three.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPowerOfThree(int n) {
         if(n<1) return false;
-        double base3 = (double)log(n)/log(3);
+        double base3 = log(n)/log(3);
         return abs(base3 - round(base3)) < 1e-10;
     }
 };
